+++ src/lib/ast/features/transform.ts
@@ -1,5 +1,4 @@
 import type {
-    AngleToken,
     AstAtRule,
     AstDeclaration,
     AstNode,
@@ -12,10 +11,12 @@ import {EnumToken} from "../types.ts";
 import {consumeWhitespace} from "../../validation/utils/index.ts";
 import {compute} from "../transform/compute.ts";
 import {filterValues, renderToken} from "../../renderer/index.ts";
-import {eqMatrix} from "../transform/minify.ts";
+import {eqMatrix, minifyTransformFunctions} from "../transform/minify.ts";
 import {FeatureWalkMode} from "./type.ts";
 
 export class TransformCssFeature {
+
+    public accept: Set<EnumToken> = new Set([EnumToken.RuleNodeType, EnumToken.KeyFramesRuleNodeType]);
 
     get ordering(): number {
         return 4;
@@ -51,47 +52,32 @@ export class TransformCssFeature {
             // @ts-ignore
             node = ast.chi[i] as AstNode | AstDeclaration;
 
-            if (
-                node.typ != EnumToken.DeclarationNodeType || !(node as AstDeclaration).nam.match(/^(-[a-z]+-)?transform$/)) {
+            if (node.typ != EnumToken.DeclarationNodeType || !(node as AstDeclaration).nam.match(/^(-[a-z]+-)?transform$/)) {
 
                 continue;
             }
 
-            const children: Token[] = (node as AstDeclaration).val.reduce((acc: Token[], curr: Token): Token[] => {
+            const children: Token[] = [];
+            for (const child of (node as AstDeclaration).val as Token[]) {
 
-
-                if (curr.typ == EnumToken.FunctionTokenType && 'skew' == curr.val.toLowerCase()) {
-
-                    if (
-                        (curr as FunctionToken).chi.length == 3) {
-
-                        if (((curr as FunctionToken).chi[2] as AngleToken).val == 0) {
-
-                            (curr as FunctionToken).chi.length = 1;
-                            (curr as FunctionToken).val = 'skew';
-                        } else if (((curr as FunctionToken).chi[0] as AngleToken).val == 0) {
-
-                            (curr as FunctionToken).chi = [(curr as FunctionToken).chi[2]];
-                            (curr as FunctionToken).val = 'skewY';
-                        }
-                    }
-                }
-
-                acc.push(curr);
-
-                return acc;
-            }, [] as Token[]);
+                children.push(child.typ == EnumToken.FunctionTokenType ? minifyTransformFunctions(child) : child);
+            }
 
             consumeWhitespace(children);
 
-            let {matrix, cumulative, minified} = compute(children as Token[]) ?? {};
+            let {matrix, cumulative, minified} = compute(children as Token[]) ?? {
+                matrix: null,
+                cumulative: null,
+                minified: null
+            };
 
             if (matrix == null || cumulative == null || minified == null) {
 
+                (node as AstDeclaration).val = children;
                 continue;
             }
 
-            let r: Token[][] = [filterValues((node as AstDeclaration).val.slice())];
+            let r: Token[][] = [filterValues(children)];
 
             if (eqMatrix(matrix as FunctionToken, cumulative)) {