+++ src/lib/validation/at-rules/layer.ts
@@ -6,6 +6,19 @@ import {validateLayerName} from "../syntaxes/index.ts";
 
 export function validateAtRuleLayer(atRule: AstAtRule, options: ValidationOptions, root?: AstNode): ValidationSyntaxResult {
 
+    if (!Array.isArray(atRule.chi)) {
+
+        // @ts-ignore
+        return {
+            valid: SyntaxValidationResult.Drop,
+            matches: [],
+            node: atRule,
+            syntax: '@' + atRule.nam,
+            error: 'expected supports body',
+            tokens: []
+        } as ValidationSyntaxResult;
+    }
+
     // media-query-list
     if (!Array.isArray(atRule.tokens) || atRule.tokens.length == 0) {